Move and rename test files for consistency
