+++ src/__tests__/logic/mocks/files/colors.js
