+++ src/__tests__/logic/mocks/files/topLevelDeclarations.js
