+++ src/types/logic-ast/declarations/__tests__/enumeration.ts
