+++ src/types/logic-ast/declarations/__tests__/function.ts
