+++ src/types/logic-ast/declarations/__tests__/namespace.ts
