+++ src/types/logic-ast/declarations/__tests__/record.ts
