+++ src/types/logic-ast/declarations/__tests__/variable.ts
