+++ src/types/logic-ast/expressions/__tests__/member.ts
