+++ src/types/logic-ast/literals/__tests__/number.ts
