+++ src/types/logic-ast/statements/__tests__/declaration.ts
